Add singleton toaster
